Move introductions to the `syntax` package
