+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/evaluation/scope/FixedDynamicScope.kt
@@ -3,7 +3,7 @@ package com.github.cubuspl42.sigmaLang.analyzer.evaluation.scope
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.Thunk
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.Value
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.toThunk
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.introductions.Declaration
+import com.github.cubuspl42.sigmaLang.analyzer.syntax.introductions.Declaration
 
 class FixedDynamicScope(
     private val entries: Map<Declaration, Value>,
