+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/evaluation/values/BuiltinGenericFunctionConstructor.kt
@@ -1,8 +1,8 @@
 package com.github.cubuspl42.sigmaLang.analyzer.evaluation.values
 
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.AbstractionConstructor
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.AtomicExpression
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.GenericType
+import com.github.cubuspl42.sigmaLang.analyzer.syntax.expressions.AbstractionConstructorTerm
 
 abstract class BuiltinGenericFunctionConstructor : AtomicExpression() {
     override val type: GenericType by lazy {
@@ -15,7 +15,7 @@ abstract class BuiltinGenericFunctionConstructor : AtomicExpression() {
     override val valueThunk: Thunk<Value>
         get() = body.valueThunk
 
-    abstract val parameterDeclaration: AbstractionConstructor.ArgumentDeclaration
+    abstract val parameterDeclaration: AbstractionConstructorTerm.ArgumentDeclaration
 
     abstract val body: BuiltinFunctionConstructor
 }
