+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/ResolvedIntroduction.kt
@@ -1,12 +1,12 @@
 package com.github.cubuspl42.sigmaLang.analyzer.semantics
 
 import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.Identifier
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.AbstractionConstructor
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.Call
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.Expression
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.FieldRead
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.IntLiteral
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.Reference
+import com.github.cubuspl42.sigmaLang.analyzer.syntax.expressions.AbstractionConstructorTerm
 
 sealed interface ResolvedIntroduction {
     fun buildReference(): Expression
@@ -15,14 +15,14 @@ sealed interface ResolvedIntroduction {
 }
 
 sealed class ResolvedAbstractionArgument : ResolvedIntroduction {
-    abstract val argumentDeclaration: AbstractionConstructor.ArgumentDeclaration
+    abstract val argumentDeclaration: AbstractionConstructorTerm.ArgumentDeclaration
 
     final override val errors: Set<SemanticError> = emptySet()
 }
 
 class ResolvedOrderedArgument(
     val index: Long,
-    override val argumentDeclaration: AbstractionConstructor.ArgumentDeclaration,
+    override val argumentDeclaration: AbstractionConstructorTerm.ArgumentDeclaration,
 ) : ResolvedAbstractionArgument() {
     override fun buildReference(): Expression = Call(
         subjectLazy = lazy { Reference(referredDeclaration = argumentDeclaration) },
@@ -32,7 +32,7 @@ class ResolvedOrderedArgument(
 
 class ResolvedUnorderedArgument(
     val name: Identifier,
-    override val argumentDeclaration: AbstractionConstructor.ArgumentDeclaration,
+    override val argumentDeclaration: AbstractionConstructorTerm.ArgumentDeclaration,
 ) : ResolvedAbstractionArgument() {
     override fun buildReference(): Expression = FieldRead(
         subjectLazy = lazy { Reference(referredDeclaration = argumentDeclaration) },
