--- analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/introductions/Introduction.kt
@@ -1,4 +1,0 @@
-package com.github.cubuspl42.sigmaLang.analyzer.semantics.introductions
-
-interface Introduction {
-}
