+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/syntax/NamespaceEntryTerm.kt
@@ -8,7 +8,7 @@ import com.github.cubuspl42.sigmaLang.analyzer.semantics.ResolvedDefinition
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.StaticScope
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.Expression
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.asStub
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.introductions.UserVariableDefinition
+import com.github.cubuspl42.sigmaLang.analyzer.syntax.introductions.UserVariableDefinition
 
 sealed interface NamespaceEntryTerm {
     companion object {
