+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/syntax/expressions/OrderedTupleTypeConstructorTerm.kt
@@ -5,7 +5,6 @@ import com.github.cubuspl42.sigmaLang.analyzer.semantics.LeveledResolvedIntroduc
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.ResolvedOrderedArgument
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.StaticBlock
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.StaticScope
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.AbstractionConstructor
 
 interface OrderedTupleTypeConstructorTerm : TupleTypeConstructorTerm {
     interface Element {
@@ -17,7 +16,7 @@ interface OrderedTupleTypeConstructorTerm : TupleTypeConstructorTerm {
     val elements: List<Element>
 
     override fun toArgumentDeclarationBlock(
-        argumentDeclaration: AbstractionConstructor.ArgumentDeclaration,
+        argumentDeclaration: AbstractionConstructorTerm.ArgumentDeclaration,
     ): StaticBlock = StaticBlock.Fixed(
         resolvedNameByName = elements.mapIndexedNotNull { index, element ->
             element.name?.let { name ->
