+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/syntax/expressions/TupleTypeConstructorTerm.kt
@@ -1,10 +1,9 @@
 package com.github.cubuspl42.sigmaLang.analyzer.syntax.expressions
 
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.StaticBlock
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.AbstractionConstructor
 
 sealed interface TupleTypeConstructorTerm : ExpressionTerm {
     fun toArgumentDeclarationBlock(
-        argumentDeclaration: AbstractionConstructor.ArgumentDeclaration,
+        argumentDeclaration: AbstractionConstructorTerm.ArgumentDeclaration,
     ): StaticBlock
 }
