+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/syntax/introductions/Declaration.kt
@@ -1,4 +1,4 @@
-package com.github.cubuspl42.sigmaLang.analyzer.semantics.introductions
+package com.github.cubuspl42.sigmaLang.analyzer.syntax.introductions
 
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.Type
