+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/syntax/introductions/Introduction.kt
@@ -1,0 +1,4 @@
+package com.github.cubuspl42.sigmaLang.analyzer.syntax.introductions
+
+interface Introduction {
+}
