+++ analyzer/src/test/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/LocalDefinitionTests.kt
@@ -1,8 +1,8 @@
 package com.github.cubuspl42.sigmaLang.analyzer.semantics
 
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.expressions.Expression
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.introductions.TypeAnnotatedBody
-import com.github.cubuspl42.sigmaLang.analyzer.semantics.introductions.UserVariableDefinition
+import com.github.cubuspl42.sigmaLang.analyzer.syntax.introductions.TypeAnnotatedBody
+import com.github.cubuspl42.sigmaLang.analyzer.syntax.introductions.UserVariableDefinition
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.BoolType
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.IntCollectiveType
 import com.github.cubuspl42.sigmaLang.analyzer.semantics.types.SpecificType
